// src/interfaces/ICart.tsx
export interface Product {
    id: number;
    title: string;
    price: number;
    quantity: number;
    total: number;
    discountPercentage: number;
    discountedPrice: number;
}

export interface Cart {
    id: number;
    products: Product[];
    total: number;
    discountedTotal: number;
    userId: number;
    totalProducts: number;
    totalQuantity: number;
}
export interface ChartData {
    labels: string[];
    datasets: {
        label: string;
        data: number[];
        fill?: boolean | undefined;
        backgroundColor?: string | undefined;
        borderColor?: string | undefined;
    }[];
}
export interface Carts {
    carts: Cart[];
}